Rename components/WhatsNew.vue to components/index/SiteTopUpper/WhatsNew.vue
